requirements.txt
pillow
pandas
openpyxl
pytesseract
numpy
opencv-python